# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pocket AI Signals Pro API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SignalRequestExpiry = typeof SignalRequestExpiry[keyof typeof SignalRequestExpiry];


export const SignalRequestExpiry = {
  '30s': '30s',
  '1m': '1m',
  '2m': '2m',
  '5m': '5m',
} as const;

export interface SignalRequest {
  /** e.g. EUR/USD (OTC) */
  currencyPair: string;
  expiry: SignalRequestExpiry;
  userId: string;
  /** @nullable */
  currentPrice?: number | null;
}

export type BatchSignalRequestExpiry = typeof BatchSignalRequestExpiry[keyof typeof BatchSignalRequestExpiry];


export const BatchSignalRequestExpiry = {
  '30s': '30s',
  '1m': '1m',
  '2m': '2m',
  '5m': '5m',
} as const;

export interface BatchSignalRequest {
  pairs: string[];
  expiry: BatchSignalRequestExpiry;
  userId: string;
}

export type SignalAction = typeof SignalAction[keyof typeof SignalAction];


export const SignalAction = {
  BUY: 'BUY',
  SELL: 'SELL',
} as const;

export interface Signal {
  action: SignalAction;
  confidence: number;
  currencyPair: string;
  expiry: string;
  analysisExplanation: string;
  /** @nullable */
  isOtc?: boolean | null;
  /** @nullable */
  rsiStatus?: string | null;
  /** @nullable */
  trendStrength?: string | null;
}

export type SignalHistoryAction = typeof SignalHistoryAction[keyof typeof SignalHistoryAction];


export const SignalHistoryAction = {
  BUY: 'BUY',
  SELL: 'SELL',
} as const;

export type SignalHistoryResult = typeof SignalHistoryResult[keyof typeof SignalHistoryResult];


export const SignalHistoryResult = {
  PENDING: 'PENDING',
  WIN: 'WIN',
  LOSS: 'LOSS',
} as const;

export interface SignalHistory {
  _id: string;
  userId: string;
  currencyPair: string;
  action: SignalHistoryAction;
  expiry: string;
  confidence: number;
  timestamp: number;
  result: SignalHistoryResult;
  /** @nullable */
  analysisExplanation?: string | null;
  /** @nullable */
  entryPrice?: number | null;
}

export type SignalHistoryInputAction = typeof SignalHistoryInputAction[keyof typeof SignalHistoryInputAction];


export const SignalHistoryInputAction = {
  BUY: 'BUY',
  SELL: 'SELL',
} as const;

export interface SignalHistoryInput {
  userId: string;
  currencyPair: string;
  action: SignalHistoryInputAction;
  expiry: string;
  confidence: number;
  timestamp: number;
  /** @nullable */
  analysisExplanation?: string | null;
  /** @nullable */
  entryPrice?: number | null;
}

export type SignalResultUpdateResult = typeof SignalResultUpdateResult[keyof typeof SignalResultUpdateResult];


export const SignalResultUpdateResult = {
  WIN: 'WIN',
  LOSS: 'LOSS',
} as const;

export interface SignalResultUpdate {
  result: SignalResultUpdateResult;
}

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  active: 'active',
  pending: 'pending',
  blocked: 'blocked',
} as const;

export interface User {
  telegramId: string;
  username: string;
  hasAccess: boolean;
  status: UserStatus;
  createdAt: number;
  /** @nullable */
  expiresAt?: number | null;
}

export interface UserInput {
  telegramId: string;
  username: string;
  /** @nullable */
  hasAccess?: boolean | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  expiresAt?: number | null;
}

export interface AccessUpdate {
  hasAccess: boolean;
  /** @nullable */
  expiresAt?: number | null;
  /** @nullable */
  status?: string | null;
}

export interface UserStats {
  userId: string;
  total: number;
  wins: number;
  losses: number;
  pending: number;
  winRate: number;
}

export interface LeaderboardEntry {
  userId: string;
  username: string;
  winRate: number;
  total: number;
}

export type GetHistoryParams = {
userId: string;
limit?: number;
};