# lib/api-zod/src/generated/types/accessUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pocket AI Signals Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface AccessUpdate {
  hasAccess: boolean;
  /** @nullable */
  expiresAt?: number | null;
  /** @nullable */
  status?: string | null;
}