# lib/api-zod/src/generated/types/signalHistory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pocket AI Signals Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { SignalHistoryAction } from './signalHistoryAction';
import type { SignalHistoryResult } from './signalHistoryResult';

export interface SignalHistory {
  _id: string;
  userId: string;
  currencyPair: string;
  action: SignalHistoryAction;
  expiry: string;
  confidence: number;
  timestamp: number;
  result: SignalHistoryResult;
  /** @nullable */
  analysisExplanation?: string | null;
  /** @nullable */
  entryPrice?: number | null;
}