# lib/api-zod/src/generated/types/signalRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pocket AI Signals Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { SignalRequestExpiry } from './signalRequestExpiry';

export interface SignalRequest {
  /** e.g. EUR/USD (OTC) */
  currencyPair: string;
  expiry: SignalRequestExpiry;
  userId: string;
  /** @nullable */
  currentPrice?: number | null;
}