# lib/api-zod/src/generated/types/signalResultUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pocket AI Signals Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { SignalResultUpdateResult } from './signalResultUpdateResult';

export interface SignalResultUpdate {
  result: SignalResultUpdateResult;
}