# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pocket AI Signals Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { UserStatus } from './userStatus';

export interface User {
  telegramId: string;
  username: string;
  hasAccess: boolean;
  status: UserStatus;
  createdAt: number;
  /** @nullable */
  expiresAt?: number | null;
}